Added an about page and a title to the index page
The about page was added by adding about.html and modifying the views.py, urls.py files.  The about page uses a simpler function call in the view rather than by class as were the other examples.  Adding the page title by variable to the index was more complicated than expected.  I ended up modifying the "context" of the view in order to add the page_title variable, which was then referenceable in the index.html file.