Added a unittest to check page title in the functional_tests.py file
This is a selenium based unittest that loads a chrome browser
with a given url localhost:8000/polls and checks the title of
the web page that loads.  Have been unable to resolve a 503
error that occurs whenever browser.quit() is called.  Not sure if
this is a legit error in this case as it is not running a selenium
grid.