Added support for adding an item via a POST
Changes in views.py, index.html for lists.  There was a bug in
index.html where the input tag was missing its name attribute. The
home_page(request) function in views.py is now returning a render
with two variables, one being the page title, with the new one being
the returned text from a POST from the item_text input control on
the form.