Enable adding items to individual lists
Added a new view "add_item", updated view_list to accept a
list_id in parameter, then to filter and render that list. Modified
URLS.PY to accept an add_item URL and updated to search explicitly
for numbers in the URL path (not as greedy).  Added NewItemTest
test cases.  Updated ListViewTest to see that only items for the
given list are displayed and to see that the expected list is
passed to the template.  Updated LIST.HTML to incorporate
receiving a list.id.