Synchronize the db.sqlite database
Somehow the LiveTestServerTestCase database was passing all
tests, but when I runserver on the live database for the functional
test, there was a persistent NOT NULL constraint failed.  I ended
up manually deleting the database file, as well as the migrations
files, then did a makemigrations and a migrate.  This seemed to
reconstruct the tables correctly (specifically the lists table), and
functional tests then pass for both databases.